Conflicts:
	JsClient/scripts/controllers/imagesController.js
	JsClient/scripts/controllers/userController.js

+++ Server/pSher.Api/Controllers/AlbumsController.cs
@@ -8,12 +8,13 @@
     using AutoMapper.QueryableExtensions;
     using Microsoft.AspNet.Identity;
     using PSher.Api.DataTransferModels.Album;
+    using PSher.Api.Validation;
     using PSher.Common.Constants;
     using PSher.Common.Extensions;
     using PSher.Services.Data.Contracts;
-
-    [EnableCors("*", "*", "*")]
+    
     [RoutePrefix("api/albums")]
+    [EnableCors("*", "*", "*")]
     public class AlbumsController : ApiController
     {
         private readonly IAlbumsService albumsService;
@@ -67,25 +68,26 @@
             var isAuthorizedAccess = this.User.Identity.IsAuthenticated;
             var currentUserId = this.User.Identity.GetUserId();
 
-            var result = await this.albumsService
-                .GetAlbumById(int.Parse(id), isAuthorizedAccess, currentUserId)
+            var resultAlbum = await this.albumsService
+                .GetAlbumById(int.Parse(id))
                 .ProjectTo<AlbumDetailsResponseModel>()
-                .ToListAsync();
+                .FirstOrDefaultAsync();
 
-            return this.Ok(result);
+            if (resultAlbum.IsPrivate 
+                && !isAuthorizedAccess 
+                && currentUserId == null
+                    || currentUserId != resultAlbum.CreatorId)
+            {
+                return this.Unauthorized();
+            }
+
+            return this.Ok(resultAlbum);
         }
 
         [Authorize]
-        [EnableCors("*", "*", "*")]
+        [ValidateModel]
         public async Task<IHttpActionResult> Post(SaveAlbumRequestModel model)
         {
-            if (!this.ModelState.IsValid || model == null)
-            {
-                return this.BadRequest(string.Format(
-                    ErrorMessages.InvalidRequestModel, 
-                    "SaveAlbumRequestModel"));
-            }
-
             var currentUserId = this.User.Identity.GetUserId();
             var tags = await this.tagsService.TagsFromCommaSeparatedValues(model.Tags);
             var images = await this.imagesService.ImagesFromCommaSeparatedIds(model.ImagesIds);
@@ -102,19 +104,23 @@
         }
 
         [Authorize]
-        [EnableCors("*", "*", "*")]
+        [ValidateModel]
         public async Task<IHttpActionResult> Put(int id, UpdateAlbumRequestModel model)
         {
-            if (!this.ModelState.IsValid || model == null)
+            var currentUserId = this.User.Identity.GetUserId();
+            var imageAuthorId = await this.albumsService.GetAlbumCreatorIdById(id);
+
+            var isCurrenUserAlbum = currentUserId == imageAuthorId;
+
+            if (!isCurrenUserAlbum)
             {
-                return this.BadRequest(string.Format(ErrorMessages.InvalidRequestModel, "UpdateAlbumRequestModel"));
+                return this.Unauthorized();
             }
 
-            var currentUserId = this.User.Identity.GetUserId();
             var tags = await this.tagsService.TagsFromCommaSeparatedValues(model.Tags);
             var images = await this.imagesService.ImagesFromCommaSeparatedIds(model.ImagesIds);
 
-            var changedAlbumId = await this.albumsService.UpdateAll(
+            var changedAlbumChangesMade = await this.albumsService.UpdateAll(
                 id,
                 model.Name,
                 currentUserId,
@@ -122,17 +128,35 @@
                 tags,
                 images);
 
-            return this.Ok(changedAlbumId);
+            if (changedAlbumChangesMade == GlobalConstants.ItemNotFoundReturnValue)
+            {
+                return this.NotFound();
+            }
+
+            return this.Ok(changedAlbumChangesMade);
         }
 
         [Authorize]
         public async Task<IHttpActionResult> Delete(int id)
         {
             var currentUserId = this.User.Identity.GetUserId();
+            var imageAuthorId = await this.albumsService.GetAlbumCreatorIdById(id);
 
-            var deletedAlbumId = await this.albumsService.Delete(id, currentUserId);
+            var isCurrenUserAlbum = currentUserId == imageAuthorId;
 
-            return this.Ok(deletedAlbumId);
+            if (!isCurrenUserAlbum)
+            {
+                return this.Unauthorized();
+            }
+
+            var deletedAlbumDeletedItems = await this.albumsService.Delete(id, currentUserId);
+
+            if (deletedAlbumDeletedItems == GlobalConstants.ItemNotFoundReturnValue)
+            {
+                return this.NotFound();
+            }
+
+            return this.Ok(deletedAlbumDeletedItems);
         }
     }
 }
